Completely Refactored Backend

+++ backend/Robot/RobotController.js
@@ -1,73 +1,125 @@
 //Import the database connection pool
 const pool = require('../Database Pool/DBConnections')
-const { v4: uuidv4 } = require('uuid');
 
-//Get the robot
+const {validateRegularID, statusCode} = require('../Utils/Formatting')
+
 const getRobot = (req,res) => {
+
     const sqlQuery = "Select * From robot"
-    pool.query(sqlQuery, (err, results)=>{
-            if(err){
-                res.status(500).json({ err: err.message})
-                return;
+
+    pool.query(sqlQuery, (error, results)=>{
+
+            if(error){
+
+                console.log("Error Fetching Robots: " + error.message)
+
+                return res.status(statusCode.INTERNAL_SERVER_ERROR).json({ error: 'Internal Server Error Fetching Robots'});
+
             }
-            res.status(200).json(results)
+
+            res.status(statusCode.OK).json(results)
+
+            return;
         }
     )
 }
 
-//Get the faulty robots
+
 const getFaultyRobot = (req,res)=>{
-    //Get faulty robot from DB
+
     const sqlQuery = "Select * From FaultyRobots"
-    pool.query(sqlQuery, (err, results)=>{
-            if(err){
-                res.status(500).json({ err: err.message})
-                return;
+
+    pool.query(sqlQuery, (error, results)=>{
+
+            if(error){
+
+                console.log("Error Fetching Faulty Robots: " + error.message)
+
+                return res.status(statusCode.INTERNAL_SERVER_ERROR).json({ error: 'Internal Server Error Fetching Faulty Robots'});
+
             }
-            res.status(200).json(results)
+
+            res.status(statusCode.OK).json(results)
+
+            return;
+
         }
     )
 }
 
-//Add a robot 
+
 const addRobot = (req, res) => {
-    //Robot data
+
     const {RobotID, CurrentLoad,RobotStatus, Maintanence, Speed,BatteryLife, EstimatedDelivery} = req.body
+
     const sqlQuery = "Insert Into Robot(RobotID, CurrentLoad, RobotStatus, Maintanence, Speed, BatteryLife, EstimatedDelivery) Values(?,?,?,?,?,?,?)"
-    pool.query(sqlQuery, [RobotID,CurrentLoad,RobotStatus,Maintanence,Speed,BatteryLife,EstimatedDelivery], (err, results)=>{
-            if(err){
-                res.status(500).json({ err: err.message})
-                return;
-            }
-            res.status(200).json({success:"True"})
+    
+    pool.query(sqlQuery, [RobotID,CurrentLoad,RobotStatus,Maintanence,Speed,BatteryLife,EstimatedDelivery], (error, results)=>{
+
+        if(error){
+
+            console.log("Error Adding Robots: " + error.message)
+
+            return res.status(statusCode.INTERNAL_SERVER_ERROR).json({ error: 'Internal Server Error Adding Robots'});
+
         }
+            res.status(statusCode.OK)
+
+            return;
+
+        }  
     )
 }
 
 const updateRobot = (req, res) => {
+
     const {RobotID, CurrentLoad,RobotStatus, Maintanence, Speed,BatteryLife, EstimatedDelivery} = req.body
+
     const sqlQuery = "Update robot set CurrentLoad = ?, RobotStatus = ?, Maintanence = ?, Speed = ?, BatteryLife = ?, EstimatedDelivery = ? Where RobotID = ?"
-    pool.query(sqlQuery, [CurrentLoad,RobotStatus,Maintanence,Speed,BatteryLife,EstimatedDelivery, RobotID], (err, results)=>{
-        if(err){
-            res.status(500).json({ err: err.message})
-            return;
+
+    pool.query(sqlQuery, [RobotID, CurrentLoad,RobotStatus, Maintanence, Speed,BatteryLife, EstimatedDelivery], (error, results)=>{
+
+        if(error){
+
+            console.log("Error Updating Robots: " + error.message)
+
+            return res.status(statusCode.INTERNAL_SERVER_ERROR).json({ error: 'Internal Server Error Updating Robots'});
+
         }
-        res.status(200).json({success:"True"})
-    }
-)
+
+            res.status(statusCode.OK)
+
+            return;
+        }  
+    )
 }
 
 const deleteRobot = (req, res) => {
     const {RobotID} = req.body
+
+    if(!validateRegularID(RobotID)){
+
+        return res.status(statusCode.BAD_REQUEST).json({error:"RobotID Is Invalid"})
+
+    }
+
     const sqlQuery = "Delete From robot Where RobotID = ?"
-    pool.query(sqlQuery, [RobotID], (err, results)=>{
-        if(err){
-            res.status(500).json({ err: err.message})
+
+    pool.query(sqlQuery, [RobotID], (error, results)=>{
+
+            if(error){
+
+                console.log("Error Deleting Robot: " + error.message)
+
+                return res.status(statusCode.INTERNAL_SERVER_ERROR).json({ error: 'Internal Server Error Deleting Robot'});
+
+            }
+
+            res.status(statusCode.OK)
+
             return;
         }
-        res.status(200).json({success:"True"})
-    }
-)
+    )
 }
 
 module.exports = {getRobot,addRobot,updateRobot,deleteRobot, getFaultyRobot}
